refactor: LottoController에 있는 함수들 역할별로 파일 분리
